first complex algorithm

+++ srcs/disorder.c
@@ -6,7 +6,7 @@
 /*   By: nahecre <nahecre@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/09 13:23:46 by nahecre           #+#    #+#             */
-/*   Updated: 2025/12/09 13:30:50 by nahecre          ###   ########.fr       */
+/*   Updated: 2025/12/10 10:38:01 by nahecre          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@ float	compute_disorder(t_stack a)
 {
 	int	mistakes;
 	int	total_pairs;
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	mistakes = 0;
 	total_pairs = 0;
